Fix CRITICAL bug: segment registers must be uint16_t not uint32_t
The process_context structure had segment registers (cs, ds, es, fs, gs, ss)
declared as uint32_t (4 bytes), but the context_switch.s assembly code treats
them as 16-bit values (2 bytes). This caused a structure layout mismatch:

- With uint32_t: offsets were cs=40, ds=44, es=48, etc.
- But assembly read: ds at offset 42, es at offset 44, etc.

This resulted in reading completely wrong values during context switching,
corrupting all registers including ESP (which showed as 0x10 instead of
0x10000000), causing Page Fault 0x2 on process execution.

Solution: Change segment registers to uint16_t to match x86 architecture
and assembly code expectations.

+++ include/process.h
@@ -25,7 +25,7 @@ struct process_context {
     uint32_t esi, edi, ebp, esp;
     uint32_t eip;
     uint32_t eflags;
-    uint32_t cs, ds, es, fs, gs, ss;
+    uint16_t cs, ds, es, fs, gs, ss;  /* Segments are 16-bit */
 } __attribute__((packed));
 
 /* Signal queue entry */
